+++ game/dota_addons/barebones/scripts/vscripts/gamemode.lua
@@ -18,6 +18,8 @@ require('libraries/physics')
 require('libraries/projectiles')
 -- This library can be used for sending panorama notifications to the UIs of players/teams/everyone
 require('libraries/notifications')
+-- This library can be used for starting customized animations on units from lua
+require('libraries/animations')
 
 -- These internal libraries set up barebones's events and processes.  Feel free to inspect them/change them if you need to.
 require('internal/gamemode')