+++ game/dota_addons/barebones/scripts/vscripts/libraries/notifications.lua
@@ -1,10 +1,10 @@
-NOTIFICATIONS_VERSION = "0.87"
+NOTIFICATIONS_VERSION = "0.88"
 
 --[[
   Sample Panorama Notifications Library by BMD
 
   Installation
-  -"require" this function inside your code in order to gain access to the Notifications class for sending notifications to players, teams, or all clients.
+  -"require" this file inside your code in order to gain access to the Notifications class for sending notifications to players, teams, or all clients.
   -Additionally, ensure that you have the barebones_hud_base.xml, barebones_hud_base.js, and barebones_hud_base.css files in your panorama content folder.
 
   Usage
@@ -28,8 +28,14 @@ NOTIFICATIONS_VERSION = "0.87"
   -For ItemImages, there is one additional mandatory parameter:
     -item:  The item name, e.g. "item_force_staff".
 
+  -Notifications can be removed from the Top/Bottom or cleared
+
   -Call the Notifications:Top, Notifications:TopToAll, or Notifications:TopToTeam to send a top-area notification to the appropriate players 
   -Call the Notifications:Bottom, Notifications:BottomToAll, or Notifications:BottomToTeam to send a bottom-area notifications to the appropriate players 
+  -Call the Notifications:ClearTop, Notifications:ClearTopFromAll, or Notifications:ClearTopFromTeam to clear all existing top-area notifications from appropriate players
+  -Call the Notifications:ClearBottom, Notifications:ClearBottomFromAll, or Notifications:ClearBottomFromTeam to clear all existing bottom-area notifications from appropriate players
+  -Call the Notifications:RemoveTop, Notifications:RemoveTopFromAll, or Notifications:RemoveTopFromTeam to remove all existing top-area notifications from appropriate players up to the provided count of notifications
+  -Call the Notifications:RemoveBottom, Notifications:RemoveBottomFromAll, or Notifications:RemoveBottomFromTeam to remove all existing bottom-area notifications from appropriate players up to the provided count of notifications
   
   Examples:
 
@@ -43,7 +49,7 @@ NOTIFICATIONS_VERSION = "0.87"
   Notifications:TopToAll({hero="npc_dota_hero_axe", imagestyle="landscape", continue=true})
   Notifications:TopToAll({hero="npc_dota_hero_axe", imagestyle="portrait", continue=true})
 
-  -- Display a generic image and then 2 ability icons and 1 item icon on the same line for 5 seconds
+  -- Display a generic image and then 2 ability icons and an item on the same line for 5 seconds
   Notifications:TopToAll({image="file://{images}/status_icons/dota_generic.psd", duration=5.0})
   Notifications:TopToAll({ability="nyx_assassin_mana_burn", continue=true})
   Notifications:TopToAll({ability="lina_fiery_soul", continue=true})
@@ -56,12 +62,82 @@ NOTIFICATIONS_VERSION = "0.87"
   Notifications:Bottom(PlayerResource:GetPlayer(0), {text="Super Size Red", duration=5, style={color="red", ["font-size"]="110px", border="10px solid blue"}})
 
 
+  -- Remove 1 bottom and 2 top notifications 2 seconds later
+  Timers:CreateTimer(2,function()
+    Notifications:RemoveTop(0, 2)
+    Notifications:RemoveBottomFromTeam(DOTA_TEAM_GOODGUYS, 1)
 
+    -- Add 1 more notification to the bottom
+    Notifications:BottomToAll({text="GREEEENNNN again", duration=9, style={color="green"}})
+  end)
+
+  -- Clear all notifications from the bottom
+  Timers:CreateTimer(7, function()
+    Notifications:ClearBottomFromAll()
+  end)
 ]]
 
 if Notifications == nil then
   Notifications = class({})
 end
+
+function Notifications:ClearTop(player)
+  Notifications:RemoveTop(player, 50)
+end
+
+function Notifications:ClearBottom(player)
+  Notifications:RemoveBottom(player, 50)
+end
+
+function Notifications:ClearTopFromAll()
+  Notifications:RemoveTopFromAll(50)
+end
+
+function Notifications:ClearBottomFromAll()
+  Notifications:RemoveBottomFromAll(50)
+end
+
+function Notifications:ClearTopFromTeam(team)
+  Notifications:RemoveTopFromTeam(team, 50)
+end
+
+function Notifications:ClearBottomFromTeam(team)
+  Notifications:RemoveBottomFromTeam(team, 50)
+end
+
+
+function Notifications:RemoveTop(player, count)
+  if type(player) == "number" then
+    player = PlayerResource:GetPlayer(player)
+  end
+
+  CustomGameEventManager:Send_ServerToPlayer(player, "top_remove_notification", {count=count} )
+end
+
+function Notifications:RemoveBottom(player, count)
+  if type(player) == "number" then
+    player = PlayerResource:GetPlayer(player)
+  end
+
+  CustomGameEventManager:Send_ServerToPlayer(player, "bottom_remove_notification", {count=count})
+end
+
+function Notifications:RemoveTopFromAll(count)
+  CustomGameEventManager:Send_ServerToAllClients("top_remove_notification", {count=count} )
+end
+
+function Notifications:RemoveBottomFromAll(count)
+  CustomGameEventManager:Send_ServerToAllClients("bottom_remove_notification", {count=count})
+end
+
+function Notifications:RemoveTopFromTeam(team, count)
+  CustomGameEventManager:Send_ServerToTeam(team, "top_remove_notification", {count=count} )
+end
+
+function Notifications:RemoveBottomFromTeam(team, count)
+  CustomGameEventManager:Send_ServerToTeam(team, "bottom_remove_notification", {count=count})
+end
+
 
 function Notifications:Top(player, table)
   if type(player) == "number" then