+++ game/dota_addons/barebones/scripts/vscripts/libraries/physics.lua
@@ -1,4 +1,4 @@
-PHYSICS_VERSION = "0.90"
+PHYSICS_VERSION = "0.91"
 
 PHYSICS_NAV_NOTHING = 0
 PHYSICS_NAV_HALT = 1
@@ -1975,7 +1975,7 @@ function Physics:PrecalculateAABox(box)
 end
 
 
-Physics:start()
+if not Physics.timers then Physics:start() end
 
 Physics:CreateColliderProfile("blocker", 
   {